training/requirements.txt
torch==2.6.0
transformers==4.56.1
sentence-transformers==5.1.0
accelerate==1.10.1
safetensors==0.6.2
numpy==2.1.2
pandas==2.3.2
scikit-learn==1.7.2
scipy==1.15.3
tqdm==4.67.1
PyYAML==6.0.2
requests==2.32.5
huggingface-hub==0.34.4
openai==1.108.1
bitsandbytes==0.47.0
peft==0.17.1
sentencepiece==0.2.1
